Comentarios Doxify
Se cambian comentarios para el doxify en formato xml
