--- src/Library/DiscordBot/Domain/BattlesList.cs
@@ -1,26 +1,0 @@
-using Discord;
-using Poke.Clases;
-
-namespace Ucu.Poo.DiscordBot.Domain;
-
-/// <summary>
-/// Esta clase representa la lista de batallas en curso.
-/// </summary>
-public class BattlesList
-{
-    private List<Battle> battles = new List<Battle>();
-
-    /// <summary>
-    /// Crea una nueva batalla entre dos jugadores.
-    /// </summary>
-    /// <param name="player1">El primer jugador.</param>
-    /// <param name="player2">El oponente.</param>
-    /// <returns>La batalla creada.</returns>
-    /// 
-    public Battle AddBattle(Trainer player1, Trainer player2)
-    {
-        var battle = new Battle(player1, player2);
-        this.battles.Add(battle);
-        return battle;
-    }
-}
