+++ src/Library/Library.csproj
@@ -16,7 +16,4 @@
     <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.0-rc.2.24473.5" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.11.1" />
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="DiscordBot\" />
-  </ItemGroup>
 </Project>
